MedGemmaFlaskAPI/requirements-api.txt
flask>=3.0.0
flask-cors>=4.0.0
transformers
torch
pillow
accelerate
bitsandbytes>=0.43.0
